pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-evaluator"
version = "0.1.0"
description = "ATS CV/Resume evaluator CLI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Your Name"}
]
dependencies = [
  "pdfminer.six>=20221105",
  "PyPDF2>=3.0.0",
  "python-docx>=1.0.0",
  "chardet>=5.0.0",
  "click>=8.1.7",
  "regex>=2024.4.28",
  "tldextract>=5.1.2"
]

[project.scripts]
ats-eval = "cv_evaluator.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "ruff>=0.6.0"
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
